+  * Cast ptrdiff_t to int for %.*s length in adnsheloex and adnslogres,
@@ -3,6 +3,8 @@
     suppress spurious autoconf error.  (Report from Mihai Ibanescu.)
+  * Cast ptrdiff_t to int for %.*s length in adnsheloex and adnslogres,
+    as is required.  (Report from Jim Meyering.)

+++ client/adnsheloex.c
@@ -168,7 +168,7 @@ static void proclog(FILE *inf, FILE *outf, int maxpending, int opts) {
       if (head->query) {
 	if (opts & OPT_DEBUG)
 	  msg("%d in queue; checking %.*s", len,
-	      head->rest-head->name, guard_null(head->name));
+	      (int)(head->rest-head->name), guard_null(head->name));
 	if (eof || len >= maxpending) {
 	  if (opts & OPT_POLL)
 	    err= adns_wait_poll(adns, &head->query, &answer, NULL);
@@ -185,7 +185,7 @@ static void proclog(FILE *inf, FILE *outf, int maxpending, int opts) {
 	if (answer->status == adns_s_ok) {
 	  const char *addr;
 	  int ok = 0;
-	  fprintf(outf, "%.*s", head->rest-head->start, head->start);
+	  fprintf(outf, "%.*s", (int)(head->rest-head->start), head->start);
 	  while(answer->nrrs--) {
 	    addr= inet_ntoa(answer->rrs.inaddr[answer->nrrs]);
 	    ok |= !strncmp(addr, head->addr, strlen(addr));
